11.4 commit

+++ qt/classes/buff.py
@@ -13,7 +13,7 @@ class Buff:
     belong: str
     buff_id: int
     buff_level: int
-    stack: int
+    stack: float
     buff_type: str
 
     name: str = ""
@@ -25,7 +25,7 @@ class Buff:
     max_stack: int = 1
     max_tick: int = 1
 
-    def __init__(self, belong: str, buff_id: int, buff_level: int, buff_type: str, stack: int = 1, **kwargs):
+    def __init__(self, belong: str, buff_id: int, buff_level: int, buff_type: str, stack: float = 1, **kwargs):
         self.belong = belong
         self.buff_id = buff_id
         self.buff_level = buff_level
